Replace an empty Misc category with Batterybar
Someone had the bright idea to add a misc category, which isn't terrible idea if we
had ANYTHING to populate it with. Batterybar was lonely and wanted some company. He has shown a
lot of progress while playing others so I have decided to let him hang out with all
of the other features in candy interface.

Change-Id: I0c972366a312e96606aa02a2005ae6c386955e60